# src/api/v1.0/swagger.json
{
  "info": {
    "version": "1.0", 
    "contact": {
      "url": "http://sdk.sphero.com/", 
      "name": "Sphero Software Engineering Support", 
      "email": "sphero-software-engineering-support@sphero.com"
    }, 
    "description": "", 
    "title": "Sphero SDK API"
  }, 
  "paths": {
...
    "/drive/rawMotors/{targetId}": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }, 
          "404": {
            "description": "Resource not found"
          }, 
          "400": {
            "description": "Bad request"
          }
        }, 
        "tags": [
          "Drive"
        ], 
        "produces": [
          "application/json"
        ], 
        "summary": "", 
        "consumes": [
          "application/json"
        ], 
        "description": ""
      }, 
      "parameters": [
        {
          "name": "targetId", 
          "format": "int32", 
          "required": true, 
          "in": "path", 
          "type": "integer", 
          "description": ""
        }, 
        {
          "schema": {
            "required": [
              "leftMode", 
              "leftSpeed", 
              "rightMode", 
              "rightSpeed"
            ], 
            "type": "object", 
            "properties": {
              "leftMode": {
                "type": "integer", 
                "format": "int32"
              }, 
              "rightMode": {
                "type": "integer", 
                "format": "int32"
              }, 
              "leftSpeed": {
                "type": "integer", 
                "format": "int32"
              }, 
              "rightSpeed": {
                "type": "integer", 
                "format": "int32"
              }
            }
          }, 
          "required": true, 
          "name": "body", 
          "in": "body"
        }
      ]
    }, 
...
    "/drive/driveWithHeading/{targetId}": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }, 
          "404": {
            "description": "Resource not found"
          }, 
          "400": {
            "description": "Bad request"
          }
        }, 
        "tags": [
          "Drive"
        ], 
        "produces": [
          "application/json"
        ], 
        "summary": "", 
        "consumes": [
          "application/json"
        ], 
        "description": ""
      }, 
      "parameters": [
        {
          "name": "targetId", 
          "format": "int32", 
          "required": true, 
          "in": "path", 
          "type": "integer", 
          "description": ""
        }, 
        {
          "schema": {
            "required": [
              "speed", 
              "heading", 
              "flags"
            ], 
            "type": "object", 
            "properties": {
              "flags": {
                "type": "integer", 
                "format": "int32"
              }, 
              "speed": {
                "type": "integer", 
                "format": "int32"
              }, 
              "heading": {
                "type": "integer", 
                "format": "int32"
              }
            }
          }, 
          "required": true, 
          "name": "body", 
          "in": "body"
        }
      ]
    }, 
...
    "/drive/resetYaw/{targetId}": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }, 
          "404": {
            "description": "Resource not found"
          }, 
          "400": {
            "description": "Bad request"
          }
        }, 
        "tags": [
          "Drive"
        ], 
        "produces": [
          "application/json"
        ], 
        "summary": "", 
        "consumes": [
          "application/json"
        ], 
        "description": ""
      }, 
      "parameters": [
        {
          "name": "targetId", 
          "format": "int32", 
          "required": true, 
          "in": "path", 
          "type": "integer", 
          "description": ""
        }
      ]
    }, 
...
    "/drive/setStabilization/{targetId}": {
      "put": {
        "responses": {
          "200": {
            "description": "OK"
          }, 
          "404": {
            "description": "Resource not found"
          }, 
          "400": {
            "description": "Bad request"
          }
        }, 
        "tags": [
          "Drive"
        ], 
        "produces": [
          "application/json"
        ], 
        "summary": "", 
        "consumes": [
          "application/json"
        ], 
        "description": ""
      }, 
      "parameters": [
        {
          "name": "targetId", 
          "format": "int32", 
          "required": true, 
          "in": "path", 
          "type": "integer", 
          "description": ""
        }, 
        {
          "schema": {
            "required": [
              "index"
            ], 
            "type": "object", 
            "properties": {
              "index": {
                "type": "integer", 
                "format": "int32"
              }
            }
          }, 
          "required": true, 
          "name": "body", 
          "in": "body"
        }
      ]
    }, 
...
  }, 
...
  "basePath": "/api/v1.0", 
  "host": "localhost:2010", 
  "swagger": "2.0"
}